WIP: Implement new module system
To increase modularity, the source files are now treated as modules.
Each module registers its own event handlers with Factorio to avoid
a mess of keeping track of which functions to call where.

The process is documented in Docs/Scriptimplementation.md.

As a test, showTakeoffDist is now its own module implemented with
the new system, other source files have yet been ported.

To fit the new system, logic/planeUtility.lua is merged with
logic/utility.lua as it is the one file which is not treated
as a module.