WIP: Improve animated shadows (smoother)
The new system for animated shadows uses LuaRendering, which allows the shadows
to be much smoother and reduces the number of sprites which must be stored.
Previously, a plane (vehicle) was created for each shadow at each scale, each
rotation.

The airborne plane does not have shadows, and a shadow is drawn below the plane
to give the illusion of climbing. The shadow's location is drawn depending on
the speed of the plane, thus handling takeoff and landing.

It currently only works with the gunship (hard coded). A system is needed to
somehow load information about each plane from the data stage to draw the shadow.