Plane data now stored in single master table
The previous method of storing information as part of the "order" property per
plane is risky as more data (shadow information) must be stored, since the
maximum length is 200 characters.  Now, the data is all kept together in a
table, serialized, and stored in prototype(s) created specifically to hold the
serialized table under its "order" property - avoiding the issue of maximum
length by splitting the serialized table across prototypes.

As a result of the change, isGroundedPlane, isAirbornePlane, getTransitionSpeed
now accepts the name of the plane, not its order.  updatePlaneShadow can now
load the shadow data dynamically per plane (previously was hard coded for
gunship).