Improve shadow transition between grounded and airborne plane
The shadow would flick forwards and back when the plane takes off. The cause
has been attributed to 3 factors:
- Using the player's position for drawing the shadow
- Shadow being drawn 1 tick behind the plane
- Plane takeoff is only processed every quarter second

To fix, the shadow is drawn using the plane's position. The shadow is moved 1
tick forwards into the future by offsetting it using the plane's velocity and
plane takeoff is checked every tick.

Also, switched to a quadratic progress instead of a linear one for the
animation to make the plane feel heavier as it is lifting off the ground.