Remove setting auto reaccelerate to keep airborne
The feature has bugs, it is possible to be kept airborne when out of fuel as it
attempts to reaccelerate the plane.
Auto reaccelerate is not very plesant as it just sets the plane to accelerate
with no limit. There are existing cruise control mods players can install in
place which behave much better.